chore : 사용하지 않는 코드를 제거한다
