+++ src/containers/minesweeper/RecordsContainer.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import { useSelector } from "react-redux";
-import Records from "../components/Records";
-import { RootState } from "../reducers";
+import Records from "../../components/Records";
+import { RootState } from "../../reducers";
 
 const RecordsContainer: FC = () => {
   const { records } = useSelector((state: RootState) => state.mineSweeper);
