+++ src/containers/minesweeper/index.ts
@@ -1,0 +1,4 @@
+export { default as ControllerContainer } from "./ControllerContainer";
+export { default as RecordsContainer } from "./RecordsContainer";
+export { default as TimeContainer } from "./TimeContainer";
+export { default as BoardContainer } from "./BoardContainer";
