+++ src/types/mineSweeper.d.ts
@@ -1,18 +1,5 @@
-type CountOfNearByMine = number;
-
-interface CellPosition {
-  row: number;
-  column: number;
-}
-
 export type Status = "INIT" | "START" | "CLEAR" | "DONE";
-export interface Cell {
-  id: string;
-  position: CellPosition;
-  status: CountOfNearByMine | "M";
-  isActive: boolean;
-  isSuspect: boolean;
-}
+
 export interface ICell {
   id: string;
   row: number;
